Node-specific exports for asyncBufferFromFile
