+++ src/utils.js
@@ -128,45 +128,6 @@ export async function asyncBufferFromUrl({ url, byteLength, requestInit, fetch: 
 }
 
 /**
- * Construct an AsyncBuffer for a local file using node fs package.
- *
- * @param {string} filename
- * @returns {Promise<AsyncBuffer>}
- */
-export async function asyncBufferFromFile(filename) {
-  const fsPackage = 'fs' // webpack no include
-  const fs = await import(fsPackage)
-  const stat = await fs.promises.stat(filename)
-  return {
-    byteLength: stat.size,
-    async slice(start, end) {
-      // read file slice
-      const readStream = fs.createReadStream(filename, { start, end })
-      return await readStreamToArrayBuffer(readStream)
-    },
-  }
-}
-
-/**
- * Convert a node ReadStream to ArrayBuffer.
- *
- * @param {import('stream').Readable} input
- * @returns {Promise<ArrayBuffer>}
- */
-function readStreamToArrayBuffer(input) {
-  return new Promise((resolve, reject) => {
-    /** @type {Buffer[]} */
-    const chunks = []
-    input.on('data', chunk => chunks.push(chunk))
-    input.on('end', () => {
-      const buffer = Buffer.concat(chunks)
-      resolve(buffer.buffer.slice(buffer.byteOffset, buffer.byteOffset + buffer.byteLength))
-    })
-    input.on('error', reject)
-  })
-}
-
-/**
  * Returns a cached layer on top of an AsyncBuffer. For caching slices of a file
  * that are read multiple times, possibly over a network.
  *
